requirements.txt
pandas
opencc-python-reimplemented
mwparserfromhell
tqdm
requests
bs4
keras
tensorflow